cache libplist

+++ pkgs/libplist/fetch.sh
@@ -1,8 +1,11 @@
 #!/bin/sh
 rm -rf "$_DESTDIR" "$_SRCDIR"
-printf "Downloading source...\n"
-curl -L -# -o src.tar.bz2 https://github.com/libimobiledevice/libplist/releases/download/2.7.0/libplist-2.7.0.tar.bz2
+ver='2.7.0'
+if [ ! -f "$_DLCACHE/libplist-$ver.tar.bz2" ] ||
+    [ "$(sha256sum "$_DLCACHE/libplist-$ver.tar.bz2" | awk '{print $1}')" != "7ac42301e896b1ebe3c654634780c82baa7cb70df8554e683ff89f7c2643eb8b" ]; then
+    printf "Downloading source...\n"
+    curl -L -# -o "$_DLCACHE/libplist-$ver.tar.bz2" "https://github.com/libimobiledevice/libplist/releases/download/$ver/libplist-$ver.tar.bz2" || exit 1
+fi
 printf "Unpacking source...\n"
-tar -C "$_TMP" -xf src.tar.bz2
-rm src.tar.bz2
+tar -C "$_TMP" -xf "$_DLCACHE/libplist-$ver.tar.bz2"
 mv "$_TMP"/libplist-* "$_SRCDIR"
